+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.10] - 2023-01-13
+
 * Use module name specified by `[package.metadata.maturin]` in [#1409](https://github.com/PyO3/maturin/pull/1409)
 
 ## [0.14.9] - 2023-01-10
@@ -797,7 +799,8 @@ points-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 2,8M | 752K | 85K
 
  * Initial Release
 
-[Unreleased]: https://github.com/pyo3/maturin/compare/v0.14.9...HEAD
+[Unreleased]: https://github.com/pyo3/maturin/compare/v0.14.10...HEAD
+[0.14.10]: https://github.com/pyo3/maturin/compare/v0.14.9...v0.14.10
 [0.14.9]: https://github.com/pyo3/maturin/compare/v0.14.8...v0.14.9
 [0.14.8]: https://github.com/pyo3/maturin/compare/v0.14.7...v0.14.8
 [0.14.7]: https://github.com/pyo3/maturin/compare/v0.14.6...v0.14.7